+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Added `\Wizapalce\SDK\Catalog\Review\ReviewService::canUserReviewProduct`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::createGroup`
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::getGroupUsers`
 
 ## 1.34.0